fix(ui): agent selection on create surfaces is a popover picker, not a wide chip row
The flat agent chip row (backend badges + names) overflowed the mobile new-session
sheet and stretched it past the viewport — the sheet is a CSS grid and the wide
row expanded the column instead of scrolling. Per the report, agent selection
should be a compact trigger that opens a secondary picker for agent + model +
reasoning effort — exactly what the chat header already has.

- Extract the chat's AgentRoutePicker (+ RouteColumn/RouteItem) out of ChatPage
  into a shared, CONTROLLED component (value + onChange). ChatPage wraps it
  (value=session, onChange=patch); the create surfaces wrap it with a draft route.
  One source of truth — no parallel agent picker.
- useNewSession now holds an agentRoute (agent/model/effort as create fields) and
  send() drops nulls into the create overrides. Replaces the single-agent chip state.
- Home + new-session sheet render the AgentRoutePicker trigger (full width) under an
  AGENT label, instead of the deleted chip-row AgentPicker.
- Fix the project chip row overflow too: min-w-0 lets it shrink in the sheet grid
  so it scrolls horizontally instead of widening the sheet.
