+++ ui/src/lib/useNewSession.ts
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { useApi } from '../context/ApiContext';
 import { useWorkbenchProjectsTree } from '../context/WorkbenchProjectsContext';
-import type { VibeAgentBrief, WorkbenchProject } from '../context/ApiContext';
+import type { VibeAgentBrief, WorkbenchProject, WorkbenchSessionCreate } from '../context/ApiContext';
 
 interface UseNewSessionOptions {
   /** Re-run the per-open reset on the rising edge — sheets pass their `open`. Default true. */
@@ -10,6 +10,19 @@ interface UseNewSessionOptions {
   /** Pre-translated copy: the hook stays i18n-free, callers pass t(...) strings. */
   loadErrorText: string;
   createFailedText: string;
+}
+
+// The agent/model/effort selection (agent route). Empty = the server default
+// (agents.default_backend). Fields allow null because the AgentRoutePicker emits
+// null to clear model/effort when switching agents; send() drops nulls before
+// creating so the create payload only carries real values.
+export interface AgentRouteSelection {
+  agent_backend?: string | null;
+  agent_name?: string | null;
+  agent_id?: string | null;
+  agent_variant?: string | null;
+  model?: string | null;
+  reasoning_effort?: string | null;
 }
 
 export interface NewSessionState {
@@ -21,14 +34,13 @@ export interface NewSessionState {
   setSelected: (id: string) => void;
   target: WorkbenchProject | null;
   needsProject: boolean;
-  // Agent (backend) selection — null = the server default (agents.default_backend).
+  // Agent route (agent + model + effort). Empty = the server default.
   agents: VibeAgentBrief[];
   defaultAgentName: string | null;
-  selectedAgent: VibeAgentBrief | null;
-  setSelectedAgent: (id: string | null) => void;
-  /** Creates a session under `target` (with the picked agent, if any) and returns the nav
-   *  target; null if it couldn't start (empty / in-flight / not loaded / no project / error).
-   *  The hook never navigates — the caller does, since the provider is mounted outside the router. */
+  agentRoute: AgentRouteSelection;
+  setAgentRoute: (patch: AgentRouteSelection) => void;
+  /** Creates a session under `target` (with the picked agent route, if any) and returns the
+   *  nav target; null if it couldn't start. The hook never navigates — the caller does. */
   send: (text: string) => Promise<{ sessionId: string; initialMessage: string } | null>;
   upsertSelectProject: (project: WorkbenchProject) => void;
 }
@@ -42,8 +54,9 @@ const sortByRecent = (list: WorkbenchProject[]) =>
 // and the mobile NewSessionSheet. A thin layer over the shared projects provider
 // (the project LIST + the create itself come from there, so a project/session
 // created here shows up in the sidebar + Projects tree). It adds the picker
-// selections (project + agent), the transient sending/error state, and target
-// resolution. Navigation + draft + the sheet's open/close lifecycle stay in the consumer.
+// selections (project + agent route), the transient sending/error state, and
+// target resolution. Navigation + draft + the sheet's open/close lifecycle stay
+// in the consumer.
 export function useNewSession({ active = true, loadErrorText, createFailedText }: UseNewSessionOptions): NewSessionState {
   const api = useApi();
   const { projects: rawProjects, projectsError, createSessionForProject, upsertProjectToTop } = useWorkbenchProjectsTree();
@@ -52,14 +65,15 @@ export function useNewSession({ active = true, loadErrorText, createFailedText }
   const [sending, setSending] = useState(false);
   const [agents, setAgents] = useState<VibeAgentBrief[]>([]);
   const [defaultAgentName, setDefaultAgentName] = useState<string | null>(null);
-  const [selectedAgentId, setSelectedAgentId] = useState<string | null>(null);
+  // The chosen agent route (agent/model/effort) as create fields; {} = default.
+  const [agentRoute, setAgentRoute] = useState<AgentRouteSelection>({});
 
   const projects = useMemo(() => (rawProjects ? sortByRecent(rawProjects) : []), [rawProjects]);
   const loaded = rawProjects !== null;
 
-  // Agents rarely change → fetch once per mount (not per sheet-open). Lets the
-  // user pick which Vibe Agent (backend) runs the session instead of always
-  // falling back to the server default.
+  // Agents rarely change → fetch once per mount (not per sheet-open). Feeds the
+  // shared AgentRoutePicker so the user can pick agent + model + effort instead
+  // of always falling back to the server default.
   useEffect(() => {
     let cancelled = false;
     api
@@ -92,7 +106,11 @@ export function useNewSession({ active = true, loadErrorText, createFailedText }
   // null or now-hidden selection still resolves a sane target.
   const target = projects.find((p) => p.id === selectedId) ?? projects[0] ?? null;
   const needsProject = loaded && !target;
-  const selectedAgent = useMemo(() => agents.find((a) => a.id === selectedAgentId) ?? null, [agents, selectedAgentId]);
+
+  const applyAgentRoute = useCallback(
+    (patch: AgentRouteSelection) => setAgentRoute((prev) => ({ ...prev, ...patch })),
+    [],
+  );
 
   const send = useCallback(
     async (text: string): Promise<{ sessionId: string; initialMessage: string } | null> => {
@@ -101,19 +119,16 @@ export function useNewSession({ active = true, loadErrorText, createFailedText }
       if (!trimmed || sending || !loaded || !target) return null;
       setSending(true);
       setError(null);
-      // null selectedAgent → omit agent fields so the server uses its default.
-      const overrides = selectedAgent
-        ? {
-            agent_id: selectedAgent.id,
-            agent_name: selectedAgent.name,
-            agent_backend: selectedAgent.backend,
-            // Match agent_variant to the backend so the session can resume its
-            // native thread (mirrors the chat AgentRoutePicker).
-            agent_variant: selectedAgent.backend,
-            model: selectedAgent.model ?? undefined,
-            reasoning_effort: selectedAgent.reasoning_effort ?? undefined,
-          }
-        : undefined;
+      // Empty agentRoute → no agent fields → the server uses its default backend.
+      // Drop null/undefined (the picker stores null to clear) so the create
+      // payload only carries the fields the user actually set.
+      const overrides: Partial<WorkbenchSessionCreate> = {};
+      if (agentRoute.agent_backend) overrides.agent_backend = agentRoute.agent_backend;
+      if (agentRoute.agent_name) overrides.agent_name = agentRoute.agent_name;
+      if (agentRoute.agent_id) overrides.agent_id = agentRoute.agent_id;
+      if (agentRoute.agent_variant) overrides.agent_variant = agentRoute.agent_variant;
+      if (agentRoute.model) overrides.model = agentRoute.model;
+      if (agentRoute.reasoning_effort) overrides.reasoning_effort = agentRoute.reasoning_effort;
       const session = await createSessionForProject(target.id, overrides);
       setSending(false);
       if (!session) {
@@ -122,7 +137,7 @@ export function useNewSession({ active = true, loadErrorText, createFailedText }
       }
       return { sessionId: session.id, initialMessage: trimmed };
     },
-    [sending, loaded, target, selectedAgent, createSessionForProject, createFailedText],
+    [sending, loaded, target, agentRoute, createSessionForProject, createFailedText],
   );
 
   const upsertSelectProject = useCallback(
@@ -148,8 +163,8 @@ export function useNewSession({ active = true, loadErrorText, createFailedText }
     needsProject,
     agents,
     defaultAgentName,
-    selectedAgent,
-    setSelectedAgent: setSelectedAgentId,
+    agentRoute,
+    setAgentRoute: applyAgentRoute,
     send,
     upsertSelectProject,
   };
